GlobalMetadata

Documentation for GlobalMetadata.

GlobalMetadata.@metadata

GlobalMetadata.@metadata!

GlobalMetadata.@metadata!

GlobalMetadata.@metadata! — Macro

@metadata!(x, k, val)

Set the value of x's global metadata associated with the key k to val.

GlobalMetadata.@metadata! — Macro

@metadata!(x[data = Dict{Symbol,Any}()])

Initializes global metadata for x to data. If data is not provided then the default is Dict{Symbol,Any}().

GlobalMetadata.@metadata — Macro

@metadata(x[, k])

Retreive metadata associated with the object id of x (objectid(x)) in the current module's global metadata. If the key k is specified only the value associated with that key is returned.